[Bugfix][Multimodal] Drop dummy DP ViT image padding
Dummy-padding local images to equalize rank work is extra ViT compute
and is not needed for all_gather. Keep output-token padding and skip
FlashInfer on 0-token queries.
